[fix] 修改輸出為旗標觸發方式而非直接觸發 [fix] 只有自己觸發的輸出才能關閉自己的輸出 [fix] 修改Mydelay_min的基準以s為單位 [add] 長按上3秒觸發強冷功能 [add] 新增[CCS]參數功能(加速冷凍) [add] 新增[CCt]參數功能

+++ INC/board_interface/board_layer.h
@@ -128,7 +128,7 @@ void set_scanflag(bool flag);
 void SMG_OPEN(uint8_t smg);
 void SMG_CLOSE(uint8_t smg);
 void Key_ReadPin(void);
-bool Mydelay_min(uint8_t flag, uint16_t min);
+bool Mydelay_min(uint8_t flag, uint32_t min);
 bool Mydelay_sec(uint8_t flag, uint16_t sec);
 bool Mydelay_ms(uint8_t flag, uint16_t ms);
 #endif /* INC_BOARD_LAYER_H_ */
